ANDROID: vfs: d_canonical_path for stacked FS
Inotify does not currently know when a filesystem is acting as a wrapper
around another fs. This means that inotify watchers will miss any
modifications to the base file, as well as any made in a separate stacked
fs that points to the same file.
d_canonical_path solves this problem by allowing the fs to map a dentry to
a path in the lower fs. Inotify can use it to find the appropriate place to
watch to be informed of all changes to a file.

[CPNOTE: 20/07/21] Lee: Requested status from Alessio via the bug

Original-Change-Id: I09563baffad1711a045e45c1bd0bd8713c2cc0b6

Test: FileObserverTest and FileObserverTestLegacyPath on Cuttlefish
Bug: 171780975
Signed-off-by: Daniel Rosenberg <drosen@google.com>
[astrachan: Folded 34df4102216e ("ANDROID: fsnotify: Notify lower fs of
            open") into this patch]
Signed-off-by: Alistair Strachan <astrachan@google.com>
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Ifb6719eaa466ab9cc98ab4b02866de16f7c73cbf
[maennich: folded into this patch:
           ffea51ec1f7a ("ANDROID: vfs: d_canonical_path cleanup #if 0")]
Signed-off-by: Matthias Maennich <maennich@google.com>
Signed-off-by: Lee Jones <joneslee@google.com>