ANDROID: scsi: ufs: allow overriding the blk_crypto_profile
Add UFSHCD_ANDROID_QUIRK_CUSTOM_CRYPTO_PROFILE which tells ufshcd-core
that the host controller supports inline encryption, but it needs to
initialize the crypto capabilities in a nonstandard way and/or it needs
to override blk_crypto_ll_ops.  If enabled, the standard code won't
initialize the blk_crypto_profile; ufs_hba_variant_ops::init() must do
it instead.

Note that it is permitted that the blk_crypto_profile has no keyslots.

This is needed for FMP support, as well as for wrapped key support.

This is a reworked version of a patch which was temporarily reverted by
https://android-review.googlesource.com/c/kernel/common/+/1867368, and
which originated from several ANDROID patches that were consolidated by
https://android-review.googlesource.com/c/kernel/common-patches/+/1508579.

[CPNOTE: 2023-08-04] Eric: blocked on upstreaming of support for
  hardware that needs this, either (a) Exynos/Pixel FMP or (b) wrapped
  key support for Qualcomm ICE, which in turn are waiting for an
  upstream-ready development platform that actually has the hardware

Bug: 162257402
Bug: 160883801
Change-Id: I556a68cd2d11bd5a7353fefdc31920475a5e7425
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Lee Jones <joneslee@google.com>