TODO: LEE: ANDROID: Revert "usb: typec: tcpm: not sink vbus if operational current is 0mA"
This reverts commit 803b1c8a0cea58cccde16eba31d285956f4c920c.

Prevents Pixel from charging. Revert while we investigate the issue.

[CPNOTE: 24/01/29] Lee: Greg reverted some of this, but leaves some diff

Test: adb shell cat /sys/class/power_supply/battery/status
Bug: 307718635
Change-Id: I9e8c6c62aac25cf41ac37c1d6bda696d8ecaf4aa
Signed-off-by: André Draszik <draszik@google.com>

ANDROID: bring back typec charger changes.

This file was originally touched to revert an upstream change that broke
the Pixel from charging, but it has now been resolved properly upstream
and the real fix will come in through the USB tree correctly.  Revert
the out-of-tree changes to allow the merge to happen properly.

Bug: 307718635
Bug: 298207935
Cc: André Draszik <draszik@google.com>
Change-Id: Icaba3965ad54d2eecf5a851520613b35770dccee
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>